# src/main/resources/swagger/gestaoobrasapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gest\u00e3o Obra API",
    "description": "API para gest\u00e3o de obras e controle de usu\u00e1rios.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/materiais/categoria": {
      "post": {
        "tags": [
          "Materiais"
        ],
        "summary": "Cadastrar Categoria de material",
        "operationId": "criarCategoriaMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaMaterialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria de material cadastrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaMaterialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Recurso relacionado n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Materiais"
        ],
        "summary": "Listar categoria de materials",
        "operationId": "listarCategoriasMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categoria de materials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoriaMaterialResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/materiais/categoria/{id}": {
      "get": {
        "tags": [
          "Materiais"
        ],
        "summary": "Buscar categoria de material por ID",
        "operationId": "buscarCategoriaMaterialPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria de material encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaMaterialResponse"
                }
              }
            }
          },
          "404": {
            "description": "Categoria de material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Materiais"
        ],
        "summary": "Atualizar categoria de material",
        "operationId": "atualizarCategoriaMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaMaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria de material atualizado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaMaterialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Categoria de material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Materiais"
        ],
        "summary": "Excluir categoria de material",
        "operationId": "excluirCategoriaMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Categoria de material exclu\u00eddo(a)"
          },
          "404": {
            "description": "Categoria de material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/materiais": {
      "post": {
        "tags": [
          "Materiais"
        ],
        "summary": "Cadastrar Material",
        "operationId": "criarMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Material cadastrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Recurso relacionado n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Materiais"
        ],
        "summary": "Listar materials",
        "operationId": "listarMateriais",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de materials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/materiais/{id}": {
      "get": {
        "tags": [
          "Materiais"
        ],
        "summary": "Buscar material por ID",
        "operationId": "buscarMaterialPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Material encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "404": {
            "description": "Material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Materiais"
        ],
        "summary": "Atualizar material",
        "operationId": "atualizarMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Material atualizado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Materiais"
        ],
        "summary": "Excluir material",
        "operationId": "excluirMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Material exclu\u00eddo(a)"
          },
          "404": {
            "description": "Material n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "CategoriaMaterialRequest": {
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 50
          },
          "nome": {
            "type": "string",
            "maxLength": 100
          }
        }
      },
      "CategoriaMaterialResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MaterialRequest": {
        "type": "object",
        "required": [
          "idCategoria",
          "idFornecedor",
          "codigo",
          "nome"
        ],
        "properties": {
          "idCategoria": {
            "type": "integer",
            "format": "int64"
          },
          "idFornecedor": {
            "type": "integer",
            "format": "int64"
          },
          "codigo": {
            "type": "string",
            "maxLength": 50
          },
          "nome": {
            "type": "string",
            "maxLength": 150
          }
        }
      },
      "MaterialResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idCategoria": {
            "type": "integer",
            "format": "int64"
          },
          "nomeCategoria": {
            "type": "string"
          },
          "idFornecedor": {
            "type": "integer",
            "format": "int64"
          },
          "nomeFornecedor": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}